ajustes driver

+++ FleetManagement.Application/Services/CompanyService.cs
@@ -1,6 +1,7 @@
 ﻿using FleetManagement.Application.DTOs.Companies;
 using FleetManagement.Application.Interfaces;
 using FleetManagement.Domain.Entities;
+using Mapster;
 
 namespace FleetManagement.Application.Services;
 
@@ -13,36 +14,51 @@ public class CompanyService : ICompanyService
         _companyRepository = companyRepository;
     }
 
-    public async Task<List<Company>> GetAllAsync()
+    public async Task<List<CompanyDto>> GetAllAsync()
     {
-        return await _companyRepository.GetAllAsync();
+        var companies = await _companyRepository.GetAllAsync();
+        return companies.Adapt<List<CompanyDto>>();
     }
 
-    public async Task<Company?> GetByIdAsync(Guid id)
+    public async Task<CompanyDto?> GetByIdAsync(Guid id)
     {
-        return await _companyRepository.GetByIdAsync(id);
+        var company = await _companyRepository.GetByIdAsync(id);
+        return company?.Adapt<CompanyDto>();
     }
 
-   public async Task CreateAsync(Company company)
-{
-    company.Id = Guid.NewGuid();
-    company.CreatedAt = DateTime.UtcNow;
-
-    await _companyRepository.AddAsync(company);
-}
-
-    public async Task UpdateAsync(Company company)
+    public async Task<CompanyDto> CreateAsync(CompanyDto dto)
     {
-        await _companyRepository.UpdateAsync(company);
+        var company = dto.Adapt<Company>();
+        company.Id = Guid.NewGuid();
+        company.CreatedAt = DateTime.UtcNow;
+
+        await _companyRepository.AddAsync(company);
+
+        return company.Adapt<CompanyDto>();
+    }
+
+    public async Task UpdateAsync(CompanyDto dto)
+    {
+        if (!dto.Id.HasValue)
+            throw new ArgumentException("Id es requerido para actualizar");
+
+        var existing = await _companyRepository.GetByIdAsync(dto.Id.Value);
+        if (existing == null)
+            throw new Exception("Company no encontrado");
+
+        existing.Name = dto.Name;
+        existing.Nit = dto.Nit;
+        existing.Address = dto.Address;
+
+        await _companyRepository.UpdateAsync(existing);
     }
 
     public async Task DeleteAsync(Guid id)
     {
-        await _companyRepository.DeleteAsync(id);
-    }
+        var existing = await _companyRepository.GetByIdAsync(id);
+        if (existing == null)
+            throw new Exception("Company no encontrado");
 
-    public Task CreateAsync(CompanyDto dto)
-    {
-        throw new NotImplementedException();
+        await _companyRepository.DeleteAsync(id);
     }
 }
